Rename SvnPanel into PanelSvg and update accordingly who is using that, as discussed in api reviews
svn path=/trunk/KDE/kdebase/workspace/libs/plasma/; revision=797663

+++ svg.cpp
@@ -361,7 +361,7 @@ QSize Svg::size() const
     return d->size.toSize();
 }
 
-void Svg::resize( int width, int height )
+void Svg::resize( qreal width, qreal height )
 {
     resize( QSize( width, height ) );
 }
